separate xps from trait to improve error handling
